+++ include/bipartite_matching.hpp
@@ -6,15 +6,15 @@
 // (k,match) を返す
 // k: 最大マッチングのペア数
 // match: 各頂点とペアになる頂点
-tuple<i64,vector<i64>> bipartite_matching(const vector<vector<i64>>& g) {
-    i64 n = SIZE(g);
-    vector<i64> match(n, -1);
+tuple<Int,vector<Int>> bipartite_matching(const vector<vector<Int>>& g) {
+    Int n = SIZE(g);
+    vector<Int> match(n, -1);
 
     BoolArray visited(n);
-    auto dfs = FIX([&g,&match,&visited](auto&& self, i64 v) -> bool {
+    auto dfs = FIX([&g,&match,&visited](auto&& self, Int v) -> bool {
         visited[v] = true;
-        for(i64 to : g[v]) {
-            i64 w = match[to];
+        for(Int to : g[v]) {
+            Int w = match[to];
             if(w == -1 || (!visited[w] && self(w))) {
                 match[v]  = to;
                 match[to] = v;
@@ -24,7 +24,7 @@ tuple<i64,vector<i64>> bipartite_matching(const vector<vector<i64>>& g) {
         return false;
     });
 
-    i64 k = 0;
+    Int k = 0;
     REP(v, n) {
         if(match[v] != -1) continue;
         ALL(fill, visited, false);
