+++ include/coord_compress.hpp
@@ -1,22 +1,22 @@
 // coord_compress {{{
 
 struct CoordCompress {
-    vector<i64> xs_;
-    unordered_map<i64,i64> comp_;
+    vector<Int> xs_;
+    unordered_map<Int,Int> comp_;
     bool built_{false};
 
     CoordCompress() = default;
 
-    explicit CoordCompress(i64 cap) {
+    explicit CoordCompress(Int cap) {
         xs_.reserve(cap);
     }
 
     template<class InputIt>
     CoordCompress(InputIt first, InputIt last) : xs_(first,last) {}
 
-    i64 size() const { return SIZE(xs_); }
+    Int size() const { return SIZE(xs_); }
 
-    void insert(i64 x) {
+    void insert(Int x) {
         ASSERT_LOCAL(!built_);
         xs_.emplace_back(x);
     }
@@ -27,7 +27,7 @@ struct CoordCompress {
         xs_.insert(end(xs_), first, last);
     }
 
-    void insert(initializer_list<i64> ilist) {
+    void insert(initializer_list<Int> ilist) {
         insert(begin(ilist), end(ilist));
     }
 
@@ -44,7 +44,7 @@ struct CoordCompress {
         built_ = true;
     }
 
-    i64 comp(i64 x) const {
+    Int comp(Int x) const {
         ASSERT_LOCAL(built_);
 #pragma GCC diagnostic push
 #pragma GCC diagnostic ignored "-Wnull-dereference"
@@ -52,17 +52,17 @@ struct CoordCompress {
 #pragma GCC diagnostic pop
     }
 
-    i64 uncomp(i64 i) const {
+    Int uncomp(Int i) const {
         ASSERT_LOCAL(built_);
         return xs_[i];
     }
 
-    i64 lower_bound(i64 x) const {
+    Int lower_bound(Int x) const {
         ASSERT_LOCAL(built_);
         return distance(begin(xs_), ALL(std::lower_bound, xs_, x));
     }
 
-    i64 upper_bound(i64 x) const {
+    Int upper_bound(Int x) const {
         ASSERT_LOCAL(built_);
         return distance(begin(xs_), ALL(std::upper_bound, xs_, x));
     }
